+++ src/main/ruby/truffleruby/core/range.rb
@@ -437,26 +437,18 @@ class Range
   end
 
   def %(n)
-    step(n)
+    Truffle::RangeOperations.step_no_block(self, n)
   end
 
   private def step_internal(step_size=1, &block) # :yields: object
-
-    if !block_given? && Primitive.object_kind_of?(self.begin, Numeric) && (Primitive.nil?(self.end) || Primitive.object_kind_of?(self.end, Numeric))
-      return Enumerator::ArithmeticSequence.new(self, :step, self.begin, self.end, step_size, self.exclude_end?)
-    end
-
-    return to_enum(:step, step_size) do
-      validated_step_args = Truffle::RangeOperations.validate_step_size(self.begin, self.end, step_size)
-      Truffle::RangeOperations.step_iterations_size(self, *validated_step_args)
-    end unless block_given?
+    return Truffle::RangeOperations.step_no_block(self, step_size) unless block
 
     values = Truffle::RangeOperations.validate_step_size(self.begin, self.end, step_size)
     first = values[0]
     last = values[1]
     step_size = values[2]
 
-    return step_endless(first, step_size, &block) if Primitive.nil? last
+    return step_endless(first, step_size, &block) if Primitive.nil?(last)
 
     case first
     when Float